marks cluster/metrics endpoint as alpha
the endpoint will be under api/v1alpha path to highlight
that it is experimental and can change in non backward
compatible way.
